+++ Api/Data/ImportInterface.php
@@ -20,6 +20,11 @@ interface ImportInterface
     public const XML_INRIVER_IMPORT_PATH_BEHAVIOR = 'inriver/import/behavior';
     public const XML_INRIVER_IMPORT_PATH_DEBUG = 'inriver/import/debug';
     public const XML_INRIVER_MAX_ALLOWED_ERROR = 'inriver/import/maximum_allowed_error';
+    public const XML_INRIVER_MAX_DOWNLOAD_RETRY_ATTEMPT = 'inriver/import/maximum_download_retry_attempt';
+    public const XML_INRIVER_INITIAL_DOWNLOAD_RETRY_SLEEP = 'inriver/import/initial_download_retry_sleep';
+    public const XML_INRIVER_MAX_DOWNLOAD_IMAGES_RETRY_ATTEMPT = 'inriver/import/maximum_download_image_retry_attempt';
+    public const XML_INRIVER_INITIAL_DOWNLOAD_IMAGES_RETRY_SLEEP = 'inriver/import/initial_download_image_retry_sleep';
+    public const XML_INRIVER_FORCE_UPDATE_STATUS_ON_CREATION = 'inriver/import/force_update_status_on_creation';
 
     /**
      * Execute import